Make ColumnTarget::column_index size_t to match get_index_by_slot_id()
Fixes a size_t -> int narrowing conversion error in the brace-init at
targets.push_back({chunk->get_index_by_slot_id(...), type_desc}). The
old scalar column_index used int only to hold a -1 "no match"
sentinel; that sentinel is now represented by an empty targets
vector instead, so there's no remaining reason for a signed type.

+++ be/src/connector/hive/scanner/hdfs_scanner_json.h
@@ -31,7 +31,7 @@ public:
     // A single json key can be mapped to more than one Hive column (e.g. mapping.c1=x and
     // mapping.c2=x both point at json field "x"), so a key resolves to a list of targets.
     struct ColumnTarget {
-        int column_index;
+        size_t column_index;
         TypeDescriptor type;
     };
